unify monitoring with sentry
