+++ src/services/logger.ts
@@ -1,4 +1,5 @@
-import {sendToSentry} from './sentry';
+import * as Sentry from '@sentry/react';
+import {SeverityLevel} from '@sentry/types/types/severity';
 
 export enum LogLevel {
   debug = 'debug',
@@ -12,7 +13,29 @@ const logWithLevel = (
   msg: string,
   obj: Record<string, unknown> = {}
 ) => {
-  sendToSentry({level, msg, obj});
+  const client = Sentry.getClient();
+  if (!client) {
+    return;
+  }
+
+  const sentryLevel = mapLogLevelToSentrySeverity(level);
+
+  if (sentryLevel === LogLevel.error) {
+    Sentry.captureException(new Error(msg), {extra: obj});
+  } else {
+    Sentry.captureEvent({
+      message: msg,
+      level: sentryLevel,
+      extra: obj,
+    });
+  }
+};
+
+export const mapLogLevelToSentrySeverity = (level: LogLevel): SeverityLevel => {
+  if (level === LogLevel.warn) {
+    return 'warning';
+  }
+  return level;
 };
 
 export const logger = {
